Split out Admin funcs to admin.go
