+++ templates/menu.html
@@ -16,7 +16,7 @@
           <ul class="pure-menu-list menu-list-dropped">
             <li class="pure-menu-item">
               {{ if .AdminMode }}
-              <a class="pure-menu-link" href="/dologout">Logout</a>
+              <a class="pure-menu-link" href="/admin/dologout">Logout</a>
               <a class="pure-menu-link" href="/">Home</a>
               {{ else }}
               <a class="pure-menu-link" href="/admin">Admin</a>
